+++ index.ts
@@ -2,3 +2,5 @@ export { isPrime } from "./src/is_prime.js";
 export { integerSqrt } from "./src/integer_sqrt.js";
 export { gcd } from "./src/gcd.js";
 export { combination } from "./src/combination.js";
+export { lcm } from "./src/lcm.js";
+export { permutation } from "./src/permutation.js";